Make sha-1s arrays of u8

+++ payload/game/kart/KartObjectManager.c
@@ -36,7 +36,7 @@ static void my_KartObjectManager_createInstance(void) {
     ai_808cb550 = 70.0f * speedModFactor;
 
     u32 courseId = s_raceConfig->raceScenario.courseId;
-    const u32 *courseSha1 = SaveManager_getCourseSha1(s_saveManager, courseId);
+    const u8 *courseSha1 = SaveManager_getCourseSha1(s_saveManager, courseId);
     SpFooter_onRaceStart(courseSha1, speedModIsEnabled);
 
     s_kartObjectManager = new(sizeof(KartObjectManager));
